+++ src/main/java/mls/util/GeneratoreRn.java
@@ -9,33 +9,40 @@ public class GeneratoreRn {
     private GeneratoreCM g;
 
     /**
-     * Generatore Rn
-     *
-     * @param a
-     * @param b
-     * @param x0
+     * @param a  Parametro a del generatore congruente moltiplicativo
+     * @param b  Parametro b potenza di m = 2^b
+     * @param x0 Seme X0
      */
-    public GeneratoreRn(long a, long b, long x0) {
+    public GeneratoreRn(long a, int b, long x0) {
        this.setG(new GeneratoreCM(a, b, x0));
     }
 
-    // genera il prossimo valore usando il GeneratoreCM dividendolo per il modulo
+    /**
+     * Restituisce valore generato tra [0,1)
+     * @return Valore generato tra [0,1)
+     */
     public double getNext() {
         double v = (double) g.getX() / g.getM();
         g.getNext(); // crea nuovo x in GeneratoreCM
         return v;
     }
 
-    // genera la sequenza in base ai parametri del generatore
+    /**
+     * Restituisce una sequenza lunga 2^b-2 tra [0,1)
+     * @return Restituisce una sequenza lunga 2^b-2 tra [0,1)
+     */
     public double[] generaSequenza() {
-        double[] l = new double[(int) Math.pow(2, g.getB()- 2)];
+        double[] l = new double[(int) Math.pow(2, g.getB()-2)];
         for(int i=0; i < l.length; i++) {
             l[i] = getNext();
         }
         return l;
     }
 
-    // genera la sequenza Zn in base ai parametri del generatore e dimensione della sequenza
+    /**
+     * Restituisce una sequenza Zn lunga size
+     * @return Restituisce una sequenza Zn lunga size
+     */
     public long[] generaSequenzaZn(double d, int size) {
         long[] l = new long[size];
         for(int i=0; i < l.length; i++) {
